desafio_metaAPI/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
chromedriver-binary==98.0.4758.48.0
click==8.0.3
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.0.0
djangorestframework==3.13.1
docstring-to-markdown==0.10
Flask==2.0.2
Flask-SocketIO==5.1.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.0.1
jedi==0.18.1
jedi-language-server==0.35.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
outcome==1.1.0
packaging==21.3
parso==0.8.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pyee==8.2.2
pygls==0.11.3
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyppeteer==1.0.2
pytest==6.2.5
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
requests==2.27.1
selenium==3.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typeguard==2.13.3
typing_extensions==4.0.1
urllib3==1.26.8
websockets==10.1
Werkzeug==2.0.2
whitenoise==5.3.0
wsproto==1.0.0
zipp==3.7.0
